Translate C to Swift (#62)
Migrate the `CMicrophonePitchTracker` C module to a new `ZenPTrack`
Swift module.

When compiling the new module with `-Ounchecked` performance is within
0%-2% of the C version. Without `-Ounchecked` it has a 20%-25% overhead.

```console
$ hyperfine --warmup 30 --runs 50 \
    pitchbench-c \
    pitchbench-swift-checked \
    pitchbench-swift-unchecked
Benchmark 1: ./pitchbench-c
  Time (mean ± σ):     391.7 ms ±   0.7 ms    [User: 379.3 ms, System: 8.6 ms]
  Range (min … max):   390.7 ms … 394.1 ms    50 runs
 
Benchmark 2: ./pitchbench-swift-checked
  Time (mean ± σ):     472.1 ms ±   0.8 ms    [User: 459.5 ms, System: 8.9 ms]
  Range (min … max):   471.0 ms … 475.8 ms    50 runs
 
Benchmark 3: ./pitchbench-swift-unchecked
  Time (mean ± σ):     391.7 ms ±   0.5 ms    [User: 379.5 ms, System: 8.6 ms]
  Range (min … max):   390.9 ms … 392.9 ms    50 runs
 
Summary
  './pitchbench-c' ran
    1.00 ± 0.00 times faster than './pitchbench-swift-unchecked'
    1.21 ± 0.00 times faster than './pitchbench-swift-checked'
```

`-Ounchecked` breaks Swift's memory safety, but the fact that it's only
enabled in release mode, combined with Swift's additional safety over C
even with that flag, means we're in a safer position than the C code
we're replacing.

+++ Packages/MicrophonePitchDetector/Sources/MicrophonePitchDetector/PitchTap.swift
@@ -42,7 +42,7 @@ final class PitchTap {
         didReceiveAudio()
 
         if tracker == nil {
-            tracker = PitchTracker(sampleRate: Int32(buffer.format.sampleRate))
+            tracker = PitchTracker(sampleRate: buffer.format.sampleRate)
         }
 
         if let pitch = tracker?.getPitch(from: buffer) {
